+++ src/main/java/muzusi/application/kis/scheduler/KisStockChartScheduler.java
@@ -1,8 +1,6 @@
 package muzusi.application.kis.scheduler;
 
 import lombok.RequiredArgsConstructor;
-import muzusi.application.kis.service.KisOAuthService;
-import muzusi.application.kis.service.KisRankingService;
 import muzusi.application.kis.service.KisStockMinutesService;
 import muzusi.domain.stock.service.StockMinutesService;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -13,31 +11,10 @@ import java.time.LocalDate;
 
 @Component
 @RequiredArgsConstructor
-public class KisScheduler {
-    private final KisOAuthService kisOAuthService;
-    private final KisRankingService kisRankingService;
+public class KisStockChartScheduler {
     private final KisStockMinutesService kisStockMinutesService;
     private final StockMinutesService stockMinutesService;
-
-    @Scheduled(cron = "0 0 7 * * ?")
-    public void runIssueAccessTokenJob() {
-        kisOAuthService.saveAccessToken();
-    }
-
-    @Scheduled(cron = "0 55 8 * * ?")
-    public void runIssueWebSocketKeyJob() {
-        kisOAuthService.saveWebSocketKey();
-    }
-
-    @Schedules({
-            @Scheduled(cron = "0 0/10 9-14 * * 1-5"),
-            @Scheduled(cron = "0 0,10,20,30 15 * * 1-5")
-    })
-    public void runRankingJobAt3PM() {
-        kisRankingService.saveVolumeRank();
-        kisRankingService.saveFluctuationRank();
-    }
-
+    
     @Schedules({
             @Scheduled(cron = "0 10,20,30,40,50 9 * * 1-5"),
             @Scheduled(cron = "0 0/10 10-14 * * 1-5"),
@@ -46,7 +23,7 @@ public class KisScheduler {
     public void runSaveStockMinutesChart() throws InterruptedException {
         kisStockMinutesService.saveStockMinutesChartAndInquirePrice();
     }
-
+    
     @Scheduled(cron = "0 0 16 * * 1-5")
     public void runSaveDailyStockMinutesChartJob() {
         kisStockMinutesService.saveDailyStockMinutesChart();
